wip: 動的パラメータを設定

+++ tobas_rc_teleop/include/tobas_rc_teleop/accel_angle.hpp
@@ -48,8 +48,8 @@ private:
 
   bool maxHorizontalAccelCb(const double& p);
   bool maxVerticalAccelCb(const double& p);
-  bool maxAttitudeCb(const double& p);
-  bool maxHeadingRateCb(const double& p);
+  bool maxAttitudeCb(const long& p);
+  bool maxHeadingRateCb(const long& p);
   bool horizontalAccelExpoCb(const long& p);
   bool verticalAccelExpoCb(const long& p);
   bool attitudeExpoCb(const long& p);
