+++ tobas_rc_teleop/include/tobas_rc_teleop/accel_yaw.hpp
@@ -41,7 +41,7 @@ private:
 
   bool maxHorizontalAccelCb(const double& p);
   bool maxVerticalAccelCb(const double& p);
-  bool maxHeadingRateCb(const double& p);
+  bool maxHeadingRateCb(const long& p);
   bool horizontalAccelExpoCb(const long& p);
   bool verticalAccelExpoCb(const long& p);
   bool headingExpoCb(const long& p);
