+++ tobas_rc_teleop/include/tobas_rc_teleop/rate_throttle.hpp
@@ -34,8 +34,8 @@ private:
   // PubSub
   ros2::PublisherPtr<tobas_command_msgs::RateThrottle> cmd_pub_;
 
-  bool maxAttitudeRateCb(const double& p);
-  bool maxHeadingRateCb(const double& p);
+  bool maxAttitudeRateCb(const long& p);
+  bool maxHeadingRateCb(const long& p);
   bool attitudeExpoCb(const long& p);
   bool headingExpoCb(const long& p);
   bool throttleExpoCb(const long& p);
